fix: Fix some problems
